requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psycopg2-binary
firebase-admin
sentence-transformers
faiss-cpu
requests
python-multipart
faster-whisper